pyproject.toml
[project]
name = "skypro"
version = "17"
description = ""
authors = [
    {name = "Nargiza Imasheva",email = "imnaur@bk.ru"}
]


readme = "README.md"
requires-python = ">=3.13"
dependencies = ["requests (>=2.32.5,<3.0.0)", "python-dotenv (>=1.2.1,<2.0.0)"]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.lint.dependencies]
mypy = "^1.18.1"
flake8 = "^7.3.0"
black = "^25.1.0"
isort = "^6.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
pytest-cov = "^7.0.0"

[tool.black]
line-length = 119
exclude = ".git"

[tool.isort]
line_length = 119

[tool.mypy]
disallow_untyped_defs = true
warn_return_any = true
exclude = "venv"
